+++ ui/src/unstructured-import/utils/file-utils.ts
@@ -1,5 +1,7 @@
 import JSZip from "jszip";
 import { saveAs } from "file-saver";
+import { ImportResult } from "../types/respons-types";
+import { dataToCypher } from "./cypher-utils";
 
 const regex = /```([\w\W]*?)\n([\w\W]*?)```/g;
 
@@ -22,4 +24,9 @@ export const saveImportResult = (importResult: string) => {
   zip.generateAsync({ type: "blob" }).then(function (content) {
     saveAs(content, "import.zip");
   });
+};
+
+export const saveCypherResult = (importResult: ImportResult) => {
+  const fileName = "neo4j_unstructured_import.cypher";
+  const cypher = dataToCypher(importResult);
 };
